Fix a memory leak caused by mistakenly using newRV_inc:
Somewhere in the API I read that creating an SV creates a reference
to it (makes sense) and that means that newRV_inc will now make two
references to it. If going out of scope removes a reference, there's
still one lingering.

I spoke to Mattia Barbon about it today and he confirmed my suspicion.
I wrote a test that showed it to be true, and decided to cover that in
case anyone wonders about it.

He also explained that blessing is done for the actual data variables,
not for references. I need to review and make sure I'm not suggesting
we are blessing references. The phrasing is "we call blesS() with a
reference", but we are effectively blessing a variable it references.