Briefly introduce variadic XSUBs
This serves as a vehicle to point out difference in behavior of many
XSUBs version Perl subs. They're close, but the rigid argument count
checking is a common pitfall. This change doesn't explain variadic XSUBs
completely since it elides any mention of how to actually inspect the
parameters to such a function, and that required and optional parameters
can be mixed. That is on purpose. The '...' is a very syntactically
obvious construct that shouldn't distract the reader too much at this
stage while allowing for instructive discussion of Perl/XS
similarities/interaction. But explaining the whole argument stack
shebang here would be a major deterrent, so that's deferred until later.