[chapter00] using "you" instead of "me" for consistency with rest of text
The rest of the text talks to the reader, whereas this section talks from
the point of view of the writer.  This change improves consistency in the
chapter.

Also changed the wording a little bit: "should generall be regarded as"
needed a "something" to follow it to make sense, and I think what is meant
here is that NO_CONTEXT should generally be used, hence the alteration.