Remove "Chapter XX" from chapter headings
This information is superfluous: there is sufficient context to work out
which chapter number this is from the directory and file names.  Also, if
someone reads the tutorial on GitHub, the chapter number information is also
provided, similarly if someone clones the project and reads it offline, the
directory and file names provide sufficient context to let the reader know
which chapter is which.

This change helps us to create a complete document via processing through
LaTeX and hence to PDF which would automatically provide chapter number
information, hence it is no longer necessary in chapter files themselves.