Add a script to convert POD to PseudoPod
This processes all chapter POD files and decrements the `=head` level by
one, meaning that the top level `=head` is `=head0` which corresponds to the
"chapter" level in PseudoPod.  Adding this filter script allows the original
POD to be displayed online, and read and edited without anything complaining
that the text isn't real POD, however still allows the files to later be
processed by `Pod::PseudoPod::LaTeX` in order to generate LaTeX output which
will then be used to create PDF output of the tutorial.