Make command under "install" a list item
The items within an `install:` block are actually parts of an array, hence
they should be prefixed with a hyphen so that they are executed correctly.
This is really only necessary when multiple commands appear under the
`install:` heading, otherwise the commands get run together, which is most
often not what one wants.